API/requirements.txt
Flask
PyMongo
pytz
python-jose-cryptodome
passlib

